feat: Add VotingProvider and VotingContext with various state variables
This commit adds a new component called VotingProvider with a VotingContext that wraps around child components. It provides various state variables such as currentAccount, candidateArray, voterArray, voterLength, etc. and functions such as checkIfWalletIsConnected, connectWallet, createVoter, getAllVoterData, etc. that are used throughout the application for fetching data and interacting with the smart contract. This commit also imports necessary dependencies such as Web3Modal, ethers, ipfs-http-client, etc. and exports a function called fetchContract that fetches the smart contract and allows us to communicate with it.